can edit and delete category from admin panel

+++ frontend-modules/angular-admin/category.js
@@ -17,11 +17,15 @@ admin.controller('CategoryListController', function ($scope, $http, $rootScope, 
     });
 
     $scope.deleteCategory = function(catId){
+        if(!confirm('Delete this category and its childs?'))
+            return;
 
         $http.delete('/api/category/' + catId)
             .success(function(data){
                 if(data.result){
                     toastr.success('Successfully deleted');
+
+                    $scope.$emit('init');
                 }
             })
             .error(function(data){
@@ -37,7 +41,10 @@ admin.controller('CategoryListController', function ($scope, $http, $rootScope, 
         }
 
         $scope.editMode = true;
-        $scope.editData = angular.copy(data);
+        $scope.editData = {
+            _id: data._id,
+            name: data.name
+        };
 
         $('#editCategoryModal').modal('show');
     };
@@ -59,10 +66,12 @@ admin.controller('CategoryListController', function ($scope, $http, $rootScope, 
 
                 if (data.result) {
                     // if successful, bind success message to message
-                    $scope.categories[data.category._id] = data.category;
+                    //$scope.categories[data.category._id] = data.category;
 
-                    $timeout(function(){
-                        $scope.$apply()});
+                    toastr.success('category edited');
+                    $('#editCategoryModal').modal('hide');
+                    $scope.$emit('init');
+
                 }
             })
             .error(function (data) {
